refactor(backend/server): split routes into modules
