+++ backend/server/src/settings/mod.rs
@@ -1,0 +1,3 @@
+mod routes;
+
+pub use routes::routes;
